Add function to calculate the enciphered key size

+++ include/astrolabous/decrypt.h
@@ -6,7 +6,7 @@
 ASTROLABOUS_EXTERN void astrolabous_recover_key(
 		uint8_t *key,
 		const uint8_t *ckey,
-		uint64_t n_hash,
+		uint32_t n_hash,
 		uint64_t n_iter);
 
 #endif
